+++ src/main/java/org/sitmun/domain/application/Application.java
@@ -17,9 +17,9 @@ import org.sitmun.domain.PersistenceConstants;
 import org.sitmun.domain.application.background.ApplicationBackground;
 import org.sitmun.domain.application.parameter.ApplicationParameter;
 import org.sitmun.domain.application.territory.ApplicationTerritory;
+import org.sitmun.domain.application.tree.ApplicationTree;
 import org.sitmun.domain.cartography.permission.CartographyPermission;
 import org.sitmun.domain.role.Role;
-import org.sitmun.domain.tree.Tree;
 import org.sitmun.domain.user.User;
 import org.sitmun.infrastructure.persistence.type.basic.Http;
 import org.sitmun.infrastructure.persistence.type.codelist.CodeList;
@@ -198,15 +198,9 @@ public class Application {
   private Set<Role> availableRoles = new HashSet<>();
 
   /** Trees assigned to this application. */
-  @ManyToMany(cascade = {CascadeType.MERGE, CascadeType.PERSIST})
-  @JoinTable(
-      name = "STM_APP_TREE",
-      joinColumns =
-          @JoinColumn(name = "ATR_APPID", foreignKey = @ForeignKey(name = "STM_ATR_FK_APP")),
-      inverseJoinColumns =
-          @JoinColumn(name = "ATR_TREEID", foreignKey = @ForeignKey(name = "STM_ATR_FK_TRE")))
+  @OneToMany(mappedBy = "application", cascade = CascadeType.ALL, orphanRemoval = true)
   @Builder.Default
-  private Set<Tree> trees = new HashSet<>();
+  private Set<ApplicationTree> trees = new HashSet<>();
 
   /** Backgrounds maps. */
   @OneToMany(mappedBy = "application", cascade = CascadeType.ALL, orphanRemoval = true)